backend/requirements.txt
mysql-connector-python
python-dotenv
speechrecognition
pyttsx3
pyaudio
flask
pymysql
flask_cors